Move some omniscidb local tests to all tests.
